+++ src/index.ts
@@ -1,7 +1,10 @@
 export * from './model';
 
-export { NucleicAcidType } from './NucleicAcidType';
-export { RNASubType } from './RNASubType';
+export { NucleicAcidType } from './enums/nucleic-acid-type';
+export { RNASubType } from './enums/rna-sub-type';
+export { AminoAcidPolarity } from './enums/amino-acid-polarity';
+export { AminoAcidCharge } from './enums/amino-acid-charge';
+export { AminoAcidSideChainType } from './enums/amino-acid-side-chain-type';
 
 export {
     ValidationResult,
@@ -13,11 +16,11 @@ export {
     chain,
     unwrap,
     unwrapOr
-} from './ValidationResult';
+} from './types/validation-result';
 
 export {
     validateNucleicAcid
-} from './validation';
+} from './utils/validation';
 
 export {
     isDNA,
@@ -28,11 +31,14 @@ export {
     isValidNucleicAcid,
     convertNucleicAcid,
     convertToDNA,
-    convertToRNA
-} from './nucleic-acids';
+    convertToRNA,
+    STOP_CODON_UAA,
+    STOP_CODON_UAG,
+    STOP_CODON_UGA,
+    STOP_CODONS
+} from './utils/nucleic-acids';
 
 export {
-    AminoAcidName,
     getAminoAcidByCodon,
     RNAtoAminoAcids
-} from './amino-acids';
+} from './utils/amino-acids';