+++ src/utils/nucleic-acids.ts
@@ -5,9 +5,9 @@ import {
     DNA,
     RNA,
     InvalidNucleotidePatternError
-} from './model';
-import { NucleicAcidType } from './NucleicAcidType';
-import { RNASubType } from './RNASubType';
+} from '../model';
+import { NucleicAcidType } from '../enums/nucleic-acid-type';
+import { RNASubType } from '../enums/rna-sub-type';
 
 /**
  * helper regex for validating nucleotide patters
@@ -299,4 +299,24 @@ export const convertToDNA = (rna: RNA): DNA => {
     const sequence = rna.getSequence();
     return new DNA(sequence.replaceAll('U', 'T'));
 };
+
+/**
+ * Stop codon UAA - does not code for an amino acid
+ */
+export const STOP_CODON_UAA = 'UAA' as const;
+
+/**
+ * Stop codon UAG - does not code for an amino acid
+ */
+export const STOP_CODON_UAG = 'UAG' as const;
+
+/**
+ * Stop codon UGA - does not code for an amino acid
+ */
+export const STOP_CODON_UGA = 'UGA' as const;
+
+/**
+ * Array of all stop codons
+ */
+export const STOP_CODONS = [STOP_CODON_UAA, STOP_CODON_UAG, STOP_CODON_UGA] as const;